Moved all searching algos to one folder
Moved all searching algos to one folder
