+++ SearchingAlgorithms/jumpsearch.cpp
